set development version to 0.31.3

+++ lsql-core/src/main/java/com/w11k/lsql/converter/types/FloatConverter.java
@@ -16,7 +16,7 @@ public class FloatConverter extends Converter {
 
     @Override
     protected void setValue(LSql lSql, PreparedStatement ps, int index, Object val) throws SQLException {
-        ps.setFloat(index, (Integer) val);
+        ps.setFloat(index, (Float) val);
     }
 
     @Override
